Report time of each System.update() if Requested
Added logic to allow the Engine class to report back the times of each
system's update operation on each frame to support deeper debugging. This
would allow consumers to swap their engine.update() call to
engine.updateDebug() if they would like to capture times of each system
they've implemented.

This should not hinder performance for normal engine.update() calls so
switching to "productionMode" is as easy as swapping one line of code in
your consuming application.

I'm certain this implementation could be better, and I'd started to look
into adding processEntity averages, for IteratingSystems, but backed off
since Engine has no great way of getting access to that data. A
"cross-cutting" utility class that can time things may be a better
approach in the future, that can be injected wherever timings would make
sense, but this looks like a simple useful start.